adding object to distro
I think there's probably a missing step.  Needs more investigation, but
it looks to me that objects are put into CloudFront (via S3) with the
Distribution, but then retrieved via URLs, and http.

If this is the case then I suppose we could use httplib2 to retrieve
the items back.  We would need to store the URLs somehow, or make them
deterministic so we can reconstruct them (if such a thing is possible).